// apps/lac-web/app/(site)/(home)/_flash-sale/type.ts
export type FeatureProduct = {
  productTitle: string;
  slug: string;
  txt_description_name: string;
  txt_mfn: string;
  txt_hazardous: string;
  txt_web_direct: string;
  txt_special_shipping: string;
  groupId: string;
  productId: string;
  group_img: string;
  product_img: string;
  sku: string;
  txt_uom: string;
  txt_uom_label: string;
  is_sale: boolean;
  is_new: boolean;
  min_order_amount: string;
  order_qty_increments: string;
  brandId: string;
  brandName: string;
  categoryId: string;
  categoryName: string;
  subCategoryId: string;
  subCategoryName: string;
};

export type MappedFeaturedProduct = {
  productTitle: string;
  productDescription: string;
  mfrPartNo: string;
  isHazardous: boolean;
  isDirectlyShippedFromVendor: boolean;
  specialShipping: boolean;
  groupId: number;
  productId: string;
  slug: string;
  groupImage: string;
  productImage: string;
  productSku: string;
  unitOfMeasure: string;
  isSaleItem: boolean;
  isNewItem: boolean;
  minimumOrderQuantity: number;
  quantityByIncrements: number;
  brandId: number;
  brandName: string;
  categoryId: number;
  categoryName: string;
  subCategoryId: number;
  subCategoryName: string;
};